Refactor: QuillEditor 컴포넌트 하나로 합치기
