--- src/components/domains/courseRegister/Quill/QuillEditor.scss
@@ -1,29 +1,0 @@
-.ql-container {
-  width: 99.1rem;
-  height: 41.1rem;
-  border-bottom-left-radius: 1rem;
-  border-bottom-right-radius: 1rem;
-  border-top: none;
-  overflow: hidden;
-}
-
-.ql-editor {
-  background-color: $White;
-  font-size: 1.5rem;
-
-  strong {
-    font-weight: 700;
-  }
-
-  i {
-    font-style: italic;
-  }
-
-  em {
-    font-style: italic;
-  }
-
-  u {
-    text-decoration: underline;
-  }
-}
